configure rest api
